requirements.txt
flask
pandas
patterns
python-binance
cs50
flask_session
werkzeug